[feat] : move the format date function to helpers file
